requirements.txt
asgiref==3.2.10
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
coverage==5.3
Django==3.1.1
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
more-itertools==8.5.0
packaging==20.4
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0
